requirements.txt
python-pptx
langchain-openai
